pokemon/pyproject.toml
[project]
name = "pokemon-lakehouse"
version = "0.1.0"
description = "Pokemon data lakehouse powered by PokeAPI"
requires-python = ">=3.11"
dependencies = [
    "phlo",
    "phlo-core-plugins",
    "phlo-dagster",
    "phlo-dlt",
    "phlo-dbt",
    "phlo-iceberg",
    "phlo-nessie",
    "phlo-trino",
    "phlo-postgres",
    "phlo-minio",
    "phlo-quality",
    "dbt-trino>=1.10.0",
]

[project.optional-dependencies]
openmetadata = [
    "phlo-openmetadata",
]

[tool.uv.sources]
phlo = { path = "../../phlo", editable = true }
phlo-core-plugins = { path = "../../phlo/packages/phlo-core-plugins", editable = true }
phlo-iceberg = { path = "../../phlo/packages/phlo-iceberg", editable = true }
phlo-dlt = { path = "../../phlo/packages/phlo-dlt", editable = true }
phlo-dbt = { path = "../../phlo/packages/phlo-dbt", editable = true }
phlo-dagster = { path = "../../phlo/packages/phlo-dagster", editable = true }
phlo-postgres = { path = "../../phlo/packages/phlo-postgres", editable = true }
phlo-minio = { path = "../../phlo/packages/phlo-minio", editable = true }
phlo-nessie = { path = "../../phlo/packages/phlo-nessie", editable = true }
phlo-trino = { path = "../../phlo/packages/phlo-trino", editable = true }
phlo-openmetadata = { path = "../../phlo/packages/phlo-openmetadata", editable = true }
phlo-quality = { path = "../../phlo/packages/phlo-quality", editable = true }
